[#69] Move current board into a view of the board. Later we will add a Rank view
