+++ src/main/resources/webapp/app/common/constants.ts
@@ -1,0 +1,3 @@
+//The toolbar height is 30px
+export const TOOLBAR_HEIGHT = 30;
+
